feat: add naive player check when playing turn
This won't prevent cheating, but it provides a basic check to ensure the 'correct' player is the one calling the function (a malicious player will still be able to play turns by changing parameters, but an oblivious player/client will be faced with an exception).
Anti-cheating is beyond the scope of this project.